updated data seeding
