--- extensions/catalog/catalog-atlas-dataseed/src/main/resources/META-INF/services/com.microsoft.dagx.spi.system.ServiceExtension
@@ -1,1 +1,0 @@
-com.microsoft.dagx.catalog.atlas.dataseed.AtlasDataSeederExtension
