+++ extensions/catalog/catalog-atlas/src/main/java/com/microsoft/dagx/catalog/atlas/metadata/AtlasExtension.java
@@ -5,6 +5,7 @@
 
 package com.microsoft.dagx.catalog.atlas.metadata;
 
+import com.microsoft.dagx.spi.metadata.MetadataStore;
 import com.microsoft.dagx.spi.security.Vault;
 import com.microsoft.dagx.spi.system.ServiceExtension;
 import com.microsoft.dagx.spi.system.ServiceExtensionContext;
@@ -45,6 +46,7 @@ public class AtlasExtension implements ServiceExtension {
         var pwd = vault.resolveSecret(SECRET_ATLAS_PWD);
         var api = new AtlasApiImpl(atlasUrl, user, pwd, context.getService(OkHttpClient.class), context.getTypeManager());
         context.registerService(AtlasApi.class, api);
+        context.registerService(MetadataStore.class, new AtlasMetadataStore(api, context.getMonitor()));
         context.getMonitor().info("Initialized Atlas API extension.");
     }
