+++ extensions/dataseed/build.gradle.kts
@@ -1,0 +1,18 @@
+/*
+ * Copyright (c) Microsoft Corporation.
+ * All rights reserved.
+ */
+
+plugins {
+    `java-library`
+}
+
+
+dependencies {
+    api(project(":spi"))
+    api(project(":extensions:dataseed:dataseed-atlas"))
+    api(project(":extensions:dataseed:dataseed-policy"))
+    api(project(":extensions:dataseed:dataseed-nifi"))
+}
+
+
