+++ extensions/dataseed/dataseed-atlas/src/main/resources/META-INF/services/com.microsoft.dagx.spi.system.ServiceExtension
@@ -1,0 +1,1 @@
+com.microsoft.dagx.dataseed.atlas.AtlasDataSeederExtension
