+++ extensions/dataseed/dataseed-atlas/src/main/resources/classifications.json
