+++ extensions/dataseed/dataseed-nifi/build.gradle.kts
@@ -1,0 +1,14 @@
+/*
+ * Copyright (c) Microsoft Corporation.
+ * All rights reserved.
+ */
+
+plugins {
+    `java-library`
+}
+
+
+dependencies {
+    api(project(":spi"))
+    implementation(project(":common"))
+}
