+++ extensions/dataseed/dataseed-nifi/src/main/java/com/microsoft/dagx/dataseed/nifi/api/Bulletin.java
@@ -1,0 +1,16 @@
+/*
+ * Copyright (c) Microsoft Corporation.
+ *  All rights reserved.
+ *
+ */
+
+package com.microsoft.dagx.dataseed.nifi.api;
+
+public class Bulletin {
+    public int id;
+    public String groupId;
+    public String sourceId;
+    public String timestamp;
+    public boolean canRead;
+    public BulletinDetails bulletin;
+}
