+++ extensions/dataseed/dataseed-nifi/src/main/java/com/microsoft/dagx/dataseed/nifi/api/BulletinBoard.java
@@ -1,0 +1,15 @@
+/*
+ * Copyright (c) Microsoft Corporation.
+ *  All rights reserved.
+ *
+ */
+
+package com.microsoft.dagx.dataseed.nifi.api;
+
+import java.util.List;
+
+public class BulletinBoard {
+    public List<Bulletin> bulletins;
+    public String generated;
+}
+
