+++ extensions/dataseed/dataseed-nifi/src/main/java/com/microsoft/dagx/dataseed/nifi/api/BulletinDetails.java
@@ -1,0 +1,18 @@
+/*
+ * Copyright (c) Microsoft Corporation.
+ *  All rights reserved.
+ *
+ */
+
+package com.microsoft.dagx.dataseed.nifi.api;
+
+public class BulletinDetails {
+    public int id;
+    public String category;
+    public String groupId;
+    public String sourceId;
+    public String sourceName;
+    public String level;
+    public String message;
+    public String timestamp;
+}
