+++ extensions/dataseed/dataseed-nifi/src/main/java/com/microsoft/dagx/dataseed/nifi/api/Component.java
@@ -1,0 +1,30 @@
+/*
+ * Copyright (c) Microsoft Corporation.
+ *  All rights reserved.
+ *
+ */
+
+package com.microsoft.dagx.dataseed.nifi.api;
+
+public class Component {
+    public String id;
+    public String name;
+    public String type;
+    public String state;
+
+    public String getId() {
+        return id;
+    }
+
+    public String getName() {
+        return name;
+    }
+
+    public String getType() {
+        return type;
+    }
+
+    public String getState() {
+        return state;
+    }
+}
