+++ extensions/dataseed/dataseed-nifi/src/main/java/com/microsoft/dagx/dataseed/nifi/api/ControllerService.java
@@ -1,0 +1,26 @@
+/*
+ * Copyright (c) Microsoft Corporation.
+ *  All rights reserved.
+ *
+ */
+
+package com.microsoft.dagx.dataseed.nifi.api;
+
+public class ControllerService {
+    public String id;
+    public String uri;
+    public Component component;
+    public Revision revision;
+
+    public String getId() {
+        return id;
+    }
+
+    public String getUri() {
+        return uri;
+    }
+
+    public Component getComponent() {
+        return component;
+    }
+}
