+++ extensions/dataseed/dataseed-nifi/src/main/java/com/microsoft/dagx/dataseed/nifi/api/ControllerServiceResponse.java
@@ -1,0 +1,14 @@
+/*
+ * Copyright (c) Microsoft Corporation.
+ *  All rights reserved.
+ *
+ */
+
+package com.microsoft.dagx.dataseed.nifi.api;
+
+import java.util.List;
+
+public class ControllerServiceResponse {
+    public String currentTime;
+    public List<ControllerService> controllerServices;
+}
