+++ extensions/dataseed/dataseed-nifi/src/main/java/com/microsoft/dagx/dataseed/nifi/api/GetBulletinResponse.java
@@ -1,0 +1,11 @@
+/*
+ * Copyright (c) Microsoft Corporation.
+ *  All rights reserved.
+ *
+ */
+
+package com.microsoft.dagx.dataseed.nifi.api;
+
+public class GetBulletinResponse {
+    public BulletinBoard bulletinBoard;
+}
