+++ extensions/dataseed/dataseed-nifi/src/main/java/com/microsoft/dagx/dataseed/nifi/api/StateDto.java
@@ -1,0 +1,22 @@
+/*
+ * Copyright (c) Microsoft Corporation.
+ *  All rights reserved.
+ *
+ */
+
+package com.microsoft.dagx.dataseed.nifi.api;
+
+public class StateDto {
+    public String id;
+    public String state;
+    public boolean disconnectedNodeAcknowledged;
+
+    public StateDto(String processGroup, String state, boolean disconnectedNodeAck) {
+        id = processGroup;
+        this.state = state;
+        disconnectedNodeAcknowledged = disconnectedNodeAck;
+    }
+
+    public StateDto() {
+    }
+}
