+++ extensions/transfer/transfer-nifi/src/test/java/com/microsoft/dagx/transfer/nifi/NifiDataFlowControllerTest.java
@@ -11,10 +11,11 @@ import com.azure.storage.blob.BlobServiceClient;
 import com.azure.storage.blob.BlobServiceClientBuilder;
 import com.azure.storage.blob.models.BlobItem;
 import com.fasterxml.jackson.databind.DeserializationFeature;
-import com.microsoft.dagx.catalog.atlas.dataseed.AzureBlobFileEntityBuilder;
 import com.microsoft.dagx.catalog.atlas.metadata.AtlasApi;
 import com.microsoft.dagx.catalog.atlas.metadata.AtlasApiImpl;
 import com.microsoft.dagx.catalog.atlas.metadata.AtlasDataCatalog;
+import com.microsoft.dagx.dataseed.atlas.AzureBlobFileEntityBuilder;
+import com.microsoft.dagx.dataseed.nifi.api.NifiApiClient;
 import com.microsoft.dagx.monitor.ConsoleMonitor;
 import com.microsoft.dagx.monitor.MonitorProvider;
 import com.microsoft.dagx.schema.SchemaRegistry;
@@ -32,7 +33,6 @@ import com.microsoft.dagx.spi.types.domain.metadata.DataEntry;
 import com.microsoft.dagx.spi.types.domain.metadata.GenericDataCatalog;
 import com.microsoft.dagx.spi.types.domain.transfer.DataAddress;
 import com.microsoft.dagx.spi.types.domain.transfer.DataRequest;
-import com.microsoft.dagx.transfer.nifi.api.NifiApiClient;
 import com.microsoft.dagx.transfer.provision.azure.AzureSasToken;
 import net.jodah.failsafe.Failsafe;
 import net.jodah.failsafe.RetryPolicy;
