--- extensions/transfer/transfer-nifi/src/test/java/com/microsoft/dagx/transfer/nifi/api/Bulletin.java
@@ -1,15 +1,0 @@
-/*
- * Copyright (c) Microsoft Corporation.
- * All rights reserved.
- */
-
-package com.microsoft.dagx.transfer.nifi.api;
-
-public class Bulletin {
-    public int id;
-    public String groupId;
-    public String sourceId;
-    public String timestamp;
-    public boolean canRead;
-    public BulletinDetails bulletin;
-}
