--- extensions/transfer/transfer-nifi/src/test/java/com/microsoft/dagx/transfer/nifi/api/BulletinBoard.java
@@ -1,14 +1,0 @@
-/*
- * Copyright (c) Microsoft Corporation.
- * All rights reserved.
- */
-
-package com.microsoft.dagx.transfer.nifi.api;
-
-import java.util.List;
-
-public class BulletinBoard {
-    public List<Bulletin> bulletins;
-    public String generated;
-}
-
