--- extensions/transfer/transfer-nifi/src/test/java/com/microsoft/dagx/transfer/nifi/api/BulletinDetails.java
@@ -1,17 +1,0 @@
-/*
- * Copyright (c) Microsoft Corporation.
- * All rights reserved.
- */
-
-package com.microsoft.dagx.transfer.nifi.api;
-
-public class BulletinDetails {
-    public int id;
-    public String category;
-    public String groupId;
-    public String sourceId;
-    public String sourceName;
-    public String level;
-    public String message;
-    public String timestamp;
-}
