--- extensions/transfer/transfer-nifi/src/test/java/com/microsoft/dagx/transfer/nifi/api/ControllerService.java
@@ -1,25 +1,0 @@
-/*
- * Copyright (c) Microsoft Corporation.
- * All rights reserved.
- */
-
-package com.microsoft.dagx.transfer.nifi.api;
-
-public class ControllerService {
-    public String id;
-    public String uri;
-    public Component component;
-    public Revision revision;
-
-    public String getId() {
-        return id;
-    }
-
-    public String getUri() {
-        return uri;
-    }
-
-    public Component getComponent() {
-        return component;
-    }
-}
