--- extensions/transfer/transfer-nifi/src/test/java/com/microsoft/dagx/transfer/nifi/api/ControllerServiceResponse.java
@@ -1,13 +1,0 @@
-/*
- * Copyright (c) Microsoft Corporation.
- * All rights reserved.
- */
-
-package com.microsoft.dagx.transfer.nifi.api;
-
-import java.util.List;
-
-public class ControllerServiceResponse {
-    public String currentTime;
-    public List<ControllerService> controllerServices;
-}
