--- extensions/transfer/transfer-nifi/src/test/java/com/microsoft/dagx/transfer/nifi/api/GetBulletinResponse.java
@@ -1,10 +1,0 @@
-/*
- * Copyright (c) Microsoft Corporation.
- * All rights reserved.
- */
-
-package com.microsoft.dagx.transfer.nifi.api;
-
-public class GetBulletinResponse {
-    public BulletinBoard bulletinBoard;
-}
