--- extensions/transfer/transfer-nifi/src/test/java/com/microsoft/dagx/transfer/nifi/api/Revision.java
@@ -1,26 +1,0 @@
-/*
- * Copyright (c) Microsoft Corporation.
- * All rights reserved.
- */
-
-package com.microsoft.dagx.transfer.nifi.api;
-
-import com.fasterxml.jackson.annotation.JsonProperty;
-
-public class Revision {
-    @JsonProperty
-    public int version;
-    @JsonProperty
-    private String clientId;
-
-    public Revision(int version) {
-        this.version = version;
-    }
-
-    public Revision() {
-    }
-
-    public String getClientId() {
-        return clientId;
-    }
-}
