--- extensions/transfer/transfer-nifi/src/test/java/com/microsoft/dagx/transfer/nifi/api/StateDto.java
@@ -1,21 +1,0 @@
-/*
- * Copyright (c) Microsoft Corporation.
- * All rights reserved.
- */
-
-package com.microsoft.dagx.transfer.nifi.api;
-
-public class StateDto {
-    public String id;
-    public String state;
-    public boolean disconnectedNodeAcknowledged;
-
-    public StateDto(String processGroup, String state, boolean disconnectedNodeAck) {
-        this.id = processGroup;
-        this.state = state;
-        this.disconnectedNodeAcknowledged = disconnectedNodeAck;
-    }
-
-    public StateDto() {
-    }
-}
